- fixed guest creation bug - refactor modals
